Phase 3.6: Slim main.ts; rename cityScene→world; state→store
main.ts extractions — main was 1157 lines mixing site boot, render
loop, live polling, source classification, and theme loading. Pulled
out:
  src/scene/renderLoop.ts   (was lines 65-587 of main: startRenderLoop,
                              _applyDisplayLabel, _setPaletteColor,
                              _orientLabelsForCamera, _resizeRendererToCanvas)
  src/store/liveUpdates.ts  (setupLiveUpdates + _clampPollSeconds +
                              LiveUpdateHandle / SignatureResponse types)
  src/utils/source.ts       (_srcKind, _deriveLabel)
  src/utils/syntaxTheme.ts  (applyHljsTheme + HLJS_VERSION)
  src/utils/url.ts          (added manifestUrl + signatureUrl)

main.ts is now ~376 lines: persistence wiring, source-picker boot
flow, syntax theme subscription, and the boot IIFE. Nothing else.

cityScene → world — `cityScene.ts` owns the whole world scene (sky,
island, trees, bushes, footprint, streets, gem, plus the city of
buildings), not just the city. Renamed:
  scene/cityScene.ts        → scene/world.ts
  createCityScene           → createWorld
  CityScene type            → World
  CitySceneDiff             → WorldDiff
  PickerCityScene           → PickerWorld
  every variable + comment reference updated globally

state → store — every file in state/ touches nanostore atoms
(liveStatus, sourceContext, configDrafts, configHotReload), and
persist.ts also manipulates atoms but was misplaced in utils/. Renamed
state/ → store/ and moved utils/persist.ts → store/persist.ts so the
folder is consistent: "everything atom-touching lives here". Other
data-layer helpers (manifestStream, url, source classification) stay
in utils/ because they don't manipulate stores.

Verified: tsc clean, all 1872 web tests pass.
