+++ web/tests/scene/picker.test.ts
@@ -12,7 +12,7 @@ import type {
   FileNode,
   FileTarget,
   DirTarget,
-  PickerCityScene,
+  PickerWorld,
   Street,
 } from '@/types';
 
@@ -55,8 +55,8 @@ function makeDirTarget(opts: { path?: string; sidewalk?: object; street?: object
 }
 
 // Minimal cityScene stub with the accessors picker actually reads. Internal
-// fixtures cast to the real types so the helper satisfies PickerCityScene.
-function makeFakeCityScene(
+// fixtures cast to the real types so the helper satisfies PickerWorld.
+function makeFakeWorld(
   initialBuildings: FakeBuildingFixture[],
   initialStreets: FakeStreetFixture[]
 ) {
@@ -94,7 +94,7 @@ function makeFakeCityScene(
   }
   setSnapshot(initialBuildings, initialStreets);
 
-  const api: PickerCityScene = {
+  const api: PickerWorld = {
     getStreetPickables() {
       return Object.keys(sidewalkMap).map((p) => sidewalkMap[p]);
     },
@@ -153,7 +153,7 @@ describe('picker.pickAtCenter', () => {
     cube.userData.street = { dir: { path: 'test' }, orientation: 'X' };
     cube.userData.type = 'directory';
 
-    const mockCityScene: PickerCityScene = {
+    const mockWorld: PickerWorld = {
       getStreetPickables: () => [cube],
       getRootGem: () => null,
       getBuildingByPath: () => null,
@@ -169,7 +169,7 @@ describe('picker.pickAtCenter', () => {
     camera.lookAt(0, 0, -10);
     camera.updateMatrixWorld();
 
-    const p = createPicker({ canvas, camera, cityScene: mockCityScene });
+    const p = createPicker({ canvas, camera, cityScene: mockWorld });
     const hit = p.pickAtCenter();
     expect(hit).not.toBeNull();
     expect(hit?.object).toBe(cube);
@@ -177,7 +177,7 @@ describe('picker.pickAtCenter', () => {
   });
 
   it('returns null when nothing is in front of the camera', () => {
-    const mockCityScene: PickerCityScene = {
+    const mockWorld: PickerWorld = {
       getStreetPickables: () => [],
       getRootGem: () => null,
       getBuildingByPath: () => null,
@@ -188,7 +188,7 @@ describe('picker.pickAtCenter', () => {
       getCells: () => new Map(),
     };
     const camera = new THREE.PerspectiveCamera(45, 1, 0.1, 1000);
-    const p = createPicker({ canvas, camera, cityScene: mockCityScene });
+    const p = createPicker({ canvas, camera, cityScene: mockWorld });
     expect(p.pickAtCenter()).toBeNull();
     p.dispose();
   });
@@ -196,7 +196,7 @@ describe('picker.pickAtCenter', () => {
 
 describe('createPicker', () => {
   it('exposes hover, selection, selectionKey atoms + setters', () => {
-    const fakeScene = makeFakeCityScene([], []);
+    const fakeScene = makeFakeWorld([], []);
     const p = createPicker({ canvas, camera: FAKE_CAMERA, cityScene: fakeScene });
     expect(typeof p.hover.get).toBe('function');
     expect(typeof p.selection.get).toBe('function');
@@ -210,7 +210,7 @@ describe('createPicker', () => {
   });
 
   it('setSelection derives selectionKey for a file target', () => {
-    const fakeScene = makeFakeCityScene([], []);
+    const fakeScene = makeFakeWorld([], []);
     const p = createPicker({ canvas, camera: FAKE_CAMERA, cityScene: fakeScene });
     p.setSelection(makeFileTarget({ path: 'src/index.js' }));
     expect(p.selectionKey.get()).toEqual({ kind: NodeKind.File, path: 'src/index.js' });
@@ -218,7 +218,7 @@ describe('createPicker', () => {
   });
 
   it('setSelection derives selectionKey for a directory target', () => {
-    const fakeScene = makeFakeCityScene([], []);
+    const fakeScene = makeFakeWorld([], []);
     const p = createPicker({ canvas, camera: FAKE_CAMERA, cityScene: fakeScene });
     p.setSelection(makeDirTarget({ path: 'src/lib' }));
     expect(p.selectionKey.get()).toEqual({ kind: NodeKind.Directory, path: 'src/lib' });
@@ -226,7 +226,7 @@ describe('createPicker', () => {
   });
 
   it('setSelection(null) clears selectionKey', () => {
-    const fakeScene = makeFakeCityScene([], []);
+    const fakeScene = makeFakeWorld([], []);
     const p = createPicker({ canvas, camera: FAKE_CAMERA, cityScene: fakeScene });
     p.setSelection(makeFileTarget({ path: 'a.js' }));
     p.setSelection(null);
@@ -236,7 +236,7 @@ describe('createPicker', () => {
 
   it('selectByPath looks up a building by path and selects it', () => {
     const meshA = { name: 'meshA' };
-    const fakeScene = makeFakeCityScene([{ path: 'a.js', mesh: meshA }], []);
+    const fakeScene = makeFakeWorld([{ path: 'a.js', mesh: meshA }], []);
     const p = createPicker({ canvas, camera: FAKE_CAMERA, cityScene: fakeScene });
     p.selectByPath('a.js');
     const sel = p.selection.get();
@@ -249,7 +249,7 @@ describe('createPicker', () => {
   });
 
   it('selectByPath(missing) leaves selection alone', () => {
-    const fakeScene = makeFakeCityScene([{ path: 'a.js', mesh: {} }], []);
+    const fakeScene = makeFakeWorld([{ path: 'a.js', mesh: {} }], []);
     const p = createPicker({ canvas, camera: FAKE_CAMERA, cityScene: fakeScene });
     p.selectByPath('a.js');
     p.selectByPath('does-not-exist.js');
@@ -262,7 +262,7 @@ describe('createPicker', () => {
 
   it('cityScene rebuild re-resolves selectionKey to a fresh selection', () => {
     const oldMesh = { id: 'old' };
-    const fakeScene = makeFakeCityScene([{ path: 'a.js', mesh: oldMesh }], []);
+    const fakeScene = makeFakeWorld([{ path: 'a.js', mesh: oldMesh }], []);
     const p = createPicker({ canvas, camera: FAKE_CAMERA, cityScene: fakeScene });
     p.selectByPath('a.js');
     const before = p.selection.get();
@@ -278,7 +278,7 @@ describe('createPicker', () => {
   });
 
   it('cityScene rebuild that removes the selected path clears selection + key', () => {
-    const fakeScene = makeFakeCityScene([{ path: 'a.js', mesh: {} }], []);
+    const fakeScene = makeFakeWorld([{ path: 'a.js', mesh: {} }], []);
     const p = createPicker({ canvas, camera: FAKE_CAMERA, cityScene: fakeScene });
     p.selectByPath('a.js');
     expect(p.selection.get()).not.toBeNull();
@@ -291,7 +291,7 @@ describe('createPicker', () => {
   });
 
   it('cityScene rebuild always clears hover (transient, can dangle on disposed mesh otherwise)', () => {
-    const fakeScene = makeFakeCityScene([{ path: 'a.js', mesh: {} }], []);
+    const fakeScene = makeFakeWorld([{ path: 'a.js', mesh: {} }], []);
     const p = createPicker({ canvas, camera: FAKE_CAMERA, cityScene: fakeScene });
     p.setHover(makeFileTarget({ mesh: { id: 'old' } }));
     expect(p.hover.get()).not.toBeNull();
@@ -302,7 +302,7 @@ describe('createPicker', () => {
   });
 
   it('interpretHit returns NodeKind.Gem for a gem hit', () => {
-    const fakeScene = makeFakeCityScene([], []);
+    const fakeScene = makeFakeWorld([], []);
     const p = createPicker({ canvas, camera: FAKE_CAMERA, cityScene: fakeScene });
     const target = p.interpretHit(fakeHit({ type: NodeKind.Gem }));
     expect(target?.kind).toBe(NodeKind.Gem);
@@ -310,7 +310,7 @@ describe('createPicker', () => {
   });
 
   it('interpretHit returns null for an uninterpretable hit', () => {
-    const fakeScene = makeFakeCityScene([], []);
+    const fakeScene = makeFakeWorld([], []);
     const p = createPicker({ canvas, camera: FAKE_CAMERA, cityScene: fakeScene });
     expect(p.interpretHit(fakeHit({}))).toBeNull();
     expect(p.interpretHit(null)).toBeNull();
